Make utxo_chains a single file
